Use a more explicit version of ExtractTextPlugin, and clean up its config
The 'extract-text-webpack-plugin' was already being used at version
'2.1.0' (hence the changes in https://github.com/girder/girder/pull/1981 ),
so this commit updates the package dependency to ensure that older versions
are not used with the new option syntax.

This new version should now properly pass on options to child loaders (
https://github.com/webpack-contrib/extract-text-webpack-plugin/pull/266 ),
so the options to 'stylus-loader' can now use a clearer syntax. This also
adds a comment to explain a poorly documented 'stylus-loader' option.